refactor: fix types
