+++ src/subscribe/mod.ts
@@ -1,7 +1,32 @@
 import not from "../utilities/not/mod.ts"
 import subscribers from "../subscribers/mod.ts"
 import type {PubSubEvent} from "../types.ts"
-import {Intl} from "./types.ts"
+
+declare namespace Intl {
+	type ListType = "conjunction" | "disjunction" | "unit"
+
+	interface ListFormatOptions {
+		localeMatcher?: "lookup" | "best fit"
+		type?: ListType
+		style?: "long" | "short" | "narrow"
+	}
+
+	interface ListFormatPart {
+		type: "element" | "literal"
+		value: string
+	}
+
+	export class ListFormat {
+		constructor(locales?: string | string[], options?: ListFormatOptions)
+		format(values: string[]): string
+		formatToParts(values: string[]): ListFormatPart[]
+		supportedLocalesOf(
+			locales: string | string[],
+			options?: ListFormatOptions,
+		): string[]
+	}
+}
+
 
 const listFormatter = new Intl.ListFormat("en", {
 	style: "long",
