--- src/subscribe/types.ts
@@ -1,24 +1,0 @@
-export declare namespace Intl {
-	type ListType = "conjunction" | "disjunction" | "unit"
-
-	interface ListFormatOptions {
-		localeMatcher?: "lookup" | "best fit"
-		type?: ListType
-		style?: "long" | "short" | "narrow"
-	}
-
-	interface ListFormatPart {
-		type: "element" | "literal"
-		value: string
-	}
-
-	class ListFormat {
-		constructor(locales?: string | string[], options?: ListFormatOptions)
-		format(values: string[]): string
-		formatToParts(values: string[]): ListFormatPart[]
-		supportedLocalesOf(
-			locales: string | string[],
-			options?: ListFormatOptions,
-		): string[]
-	}
-}
