--- src/subscribers/types.ts
@@ -1,7 +1,0 @@
-import type {Topics} from "../types.ts"
-
-export type Subscribers = {
-	once?: Topics
-	always?: Topics
-}
-
